move utils out of provider directory
