+++ src/hooks/DAO/proposal/useSubmitProposal.ts
@@ -8,9 +8,9 @@ import { useProvider, useSigner } from 'wagmi';
 import { buildSafeAPIPost, encodeMultiSend } from '../../../helpers';
 import { logError } from '../../../helpers/errorLogging';
 import { useFractal } from '../../../providers/App/AppProvider';
-import { buildGnosisApiUrl } from '../../../providers/Fractal/utils';
 import { useNetworkConfg } from '../../../providers/NetworkConfig/NetworkConfigProvider';
 import { MetaTransaction, ProposalExecuteData } from '../../../types';
+import { buildGnosisApiUrl } from '../../../utils';
 import useSafeContracts from '../../safe/useSafeContracts';
 import { useSafeMultisigProposals } from '../loaders/governance/useSafeMultisigProposals';
 import { useFractalModules } from '../loaders/useFractalModules';
