+++ src/providers/Fractal/governance/hooks/useUpdateProposalState.ts
@@ -1,8 +1,8 @@
 import { BigNumber } from 'ethers';
 import { useCallback, Dispatch } from 'react';
 import { FractalGovernanceContracts } from '../../../../types';
+import { getTxProposalState } from '../../../../utils';
 import { FractalGovernanceAction, FractalGovernanceActions } from '../../../App/governance/action';
-import { getTxProposalState } from '../../utils';
 
 interface IUseUpdateProposalState {
   governanceContracts: FractalGovernanceContracts;
