+++ src/utils/currency.ts
@@ -7,3 +7,12 @@ export const formatWeiToValue = (amountInWei: string | BigNumber, decimals: numb
     maximumFractionDigits: decimals,
   }).format(value);
 };
+
+export const numberFormatted = (number: number, decimals: number, currencyId?: string) => {
+  return new Intl.NumberFormat(undefined, {
+    minimumFractionDigits: 2,
+    maximumFractionDigits: decimals,
+    style: currencyId ? 'currency' : undefined,
+    currency: currencyId,
+  }).format(number);
+};
